Fixes #22106 - Clean up the common components
